Fix advisor review nits: return type, stale Meteostat prose
- _lag_days returns int, not float
- retrieve_gap_fill_forecasts docstring named Meteostat as the historical
  source; it is the Open-Meteo archive now
- Six more stale Meteostat references in app.py, including the chart
  legend "Historical (Meteostat hourly)" and the data-load error message
  telling the user to check a service we no longer call

+++ test_data_freshness.py
@@ -53,7 +53,7 @@ MAX_LAG_DAYS = {
 }
 
 
-def _lag_days(latest) -> float:
+def _lag_days(latest) -> int:
     """Whole days between the latest datapoint and today."""
     now = pd.Timestamp.now().normalize()
     return (now - pd.Timestamp(latest).normalize()).days
